add gitignore

+++ controler_in_list_employes/supprimerlisteemployes.php
@@ -1,24 +1,28 @@
-<?php 
+<?php
 
-include 'connexion.php';
-if (isset($_GET['id'])) {
+include '../connexion.php';
+
+if (isset($_POST['user_id'])) {
     // Récupération de l'ID à supprimer
-    $id = $_GET['id'];
-
-$pdostat = $connexion->prepare('DELETE FROM users WHERE id=:id LIMIT 1 ');
-
-$pdostat->bindValue(':id' , $_GET['id'] ,PDO::PARAM_INT);
-
-$executeisOk = $pdostat->execute();
+    $id = $_POST['user_id'];
+    $redirect = $_POST['redirect'];
 
 
-if ($executeisOk) {
-    header( 'location: listEmployesAdmin.php' );
+    $pdostat = $connexion->prepare('DELETE FROM users WHERE id=:id LIMIT 1 ');
+
+    $pdostat->bindValue(':id', $id, PDO::PARAM_INT);
+
+    $executeisOk = $pdostat->execute();
+
+    if ($executeisOk) {
+        if ($redirect !== null) {
+            header('location: ../' . $redirect);
+        } else {
+            header('location: in_listClient.php');
+        }
+    } else {
+        echo  "echec de la suppression du client  ";
+    }
 } else {
-    echo  "echec de la suppression du client  " ;
-}
-}else{
     echo "Erreur : aucun ID de client fourni pour la suppression.";
 }
-
-?>
